requirements.txt
Django>=1.4
django-debug-toolbar>=1.0
pylibmc
python-memcached
